Refactor 'classes' because it isnt a command just a utility module
