+++ kalite/distributed/management/commands/retrievecontentpack.py
@@ -16,7 +16,7 @@ from kalite.contentload import settings as content_settings
 from kalite.i18n.base import lcode_to_django_lang, get_po_filepath, get_locale_path, \
     download_content_pack, update_jsi18n_file, get_subtitle_file_path as get_subtitle_path, \
     extract_content_db, get_localized_exercise_dirpath
-from kalite.updates.management.commands.classes import UpdatesStaticCommand
+from kalite.updates.management.utils import UpdatesStaticCommand
 
 logging = django_settings.LOG
